# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.DownloadCertificateResponse:
    properties:
      cert_pem:
        type: string
      domain:
        type: string
      full_chain_pem:
        type: string
      issuer_cert_pem:
        type: string
    type: object
...
  handler.RequestCertificateInput:
    properties:
      auto_renew:
        type: boolean
      dns_provider_id:
        type: integer
      domain:
        type: string
      proxy_id:
        type: integer
    required:
    - dns_provider_id
    - domain
    type: object
  handler.RequestCertificateResponse:
    properties:
      domain:
        type: string
      task_id:
        type: string
    type: object
  handler.UpdateAutoRenewInput:
    properties:
      auto_renew:
        type: boolean
    type: object
...
  util.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: FRP 管理面板后端 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FRP 管理面板 API
  version: "1.0"
paths:
...
  /api/certificates:
    get:
      consumes:
      - application/json
      description: 获取所有 SSL 证书的列表，包含证书状态、域名、过期时间等信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回证书列表
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取证书列表
      tags:
      - 证书管理
    post:
      consumes:
      - application/json
      description: 异步申请新的 SSL 证书，通过 DNS 验证方式完成域名所有权验证
      parameters:
      - description: 证书申请信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.RequestCertificateInput'
      produces:
      - application/json
      responses:
        "200":
          description: 成功提交申请，返回任务ID
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.RequestCertificateResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 申请证书
      tags:
      - 证书管理
  /api/certificates/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的 SSL 证书，删除后无法恢复
      parameters:
      - description: 证书ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: 无效的证书ID
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: 证书不存在
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 删除证书失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 删除证书
      tags:
      - 证书管理
    get:
      consumes:
      - application/json
      description: 根据证书 ID 获取单个证书的详细信息
      parameters:
      - description: 证书ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回证书详情
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 无效的证书ID
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: 证书不存在
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取证书详情
      tags:
      - 证书管理
  /api/certificates/{id}/auto-renew:
    put:
      consumes:
      - application/json
      description: 更新证书的自动续期状态，开启后系统将在证书即将过期时自动续期
      parameters:
      - description: 证书ID
        in: path
        name: id
        required: true
        type: integer
      - description: 自动续期状态
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateAutoRenewInput'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: 无效的证书ID或参数错误
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: 证书不存在
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 设置自动续期
      tags:
      - 证书管理
  /api/certificates/{id}/download:
    get:
      consumes:
      - application/json
      description: 下载证书内容，包含证书 PEM、颁发者证书 PEM 和完整证书链
      parameters:
      - description: 证书ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回证书内容
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.DownloadCertificateResponse'
              type: object
        "400":
          description: 无效的证书ID或证书状态无效
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: 证书不存在
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 下载证书
      tags:
      - 证书管理
  /api/certificates/{id}/reapply:
    post:
      consumes:
      - application/json
      description: 重新申请之前申请失败的证书，使用原有的域名和 DNS 提供商配置
      parameters:
      - description: 证书ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 重新申请成功
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: 无效的证书ID
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 重新申请证书失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 重新申请证书
      tags:
      - 证书管理
  /api/certificates/{id}/renew:
    post:
      consumes:
      - application/json
      description: 手动续期指定的 SSL 证书，重新向 ACME 服务器申请新证书
      parameters:
      - description: 证书ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 续期成功
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: 无效的证书ID
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 续期证书失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 续期证书
      tags:
      - 证书管理
  /api/certificates/active:
    get:
      consumes:
      - application/json
      description: 获取所有状态为有效的证书列表，不包含已过期或申请失败的证书
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回活跃证书列表
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "500":
          description: 查询证书失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取活跃证书
      tags:
      - 证书管理
  /api/certificates/by-domain:
    get:
      consumes:
      - application/json
      description: 根据精确域名查询对应的证书列表
      parameters:
      - description: 域名
        in: query
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回证书列表
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "400":
          description: 域名不能为空
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 查询证书失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 按域名查询证书
      tags:
      - 证书管理
  /api/certificates/expiring:
    get:
      consumes:
      - application/json
      description: 获取所有即将在 30 天内过期的证书列表，用于提醒续期
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回即将过期的证书列表
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "500":
          description: 查询证书失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 获取即将过期的证书
      tags:
      - 证书管理
  /api/certificates/match:
    get:
      consumes:
      - application/json
      description: 根据域名模式匹配证书，支持通配符证书匹配（如 *.example.com 可匹配 sub.example.com）
      parameters:
      - description: 域名
        in: query
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回匹配的证书列表
          schema:
            allOf:
            - $ref: '#/definitions/util.Response'
            - properties:
                data:
                  items:
                    type: object
                  type: array
              type: object
        "400":
          description: 域名不能为空
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: 查询证书失败
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerAuth: []
      summary: 匹配证书
      tags:
      - 证书管理
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"